// convex/plugins/wca/openapiClient/types.gen.ts
export type UserAvatar = {
  url: string
  thumb_url?: string
  is_default: boolean
}
// ...
export type CompetitionIndex = {
  id: string
  name: string
  short_display_name: string
  start_date: string
  end_date: string
  registration_open: string
  registration_close: string
  announced_at: string
  country_iso2: string
  city: string
  venue: string
  latitude_degrees: number
  longitude_degrees: number
  event_ids: Array<string>
  main_event_id: string
  competitor_limit: number
  championship_types: Array<string>
  results_submitted_at?: string
  results_posted_at?: string
  report_posted_at?: string
  report_posted_by_user?: number
  delegates?: Array<{
    id: number
    name: string
    avatar: UserAvatar
  }>
}

export type MyCompetition = {
  id: string
  name: string
  website: string
  start_date: string
  end_date: string
  registration_open: string
  url: string
  city: string
  country_iso2: string
  "results_posted?": boolean
  "report_posted?": boolean
  "visible?": boolean
  "confirmed?": boolean
  "cancelled?": boolean
  short_display_name: string
  championships: Array<string>
  registration_status?: string
}